+++ lib/domain/record/components/pill_sheet/components/record_page_pill_option.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:pilll/components/organisms/pill_sheet/pill_sheet_view_layout.dart';
-import 'package:pilll/domain/record/components/pill_sheet/components/manual_rest_duration_button.dart';
+import 'package:pilll/domain/record/components/rest_duration/manual_rest_duration_button.dart';
 import 'package:pilll/domain/record/record_page_store.dart';
 import 'package:pilll/entity/pill_sheet.dart';
 import 'package:pilll/entity/pill_sheet_group.dart';
